requirements.txt
# Core dependencies
cryptography>=41.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM Integration (choose based on preference)
openai>=1.0.0  # For OpenAI API
anthropic>=0.7.0  # For Claude API
# ollama  # Uncomment for local Ollama support

# CLI Interface
click>=8.1.0
rich>=13.0.0
prompt-toolkit>=3.0.0

# Data persistence
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0
